Extract ReplayPlot so ReplayCurve is only hero chrome.
Stop the plot+annotation sprawl by giving the dramatized climb its own module; ReplayCurve keeps freeze header, stamp, and footnote.
